backend/requirements.txt
# Core Django
Django==5.1.4                    # Latest stable
djangorestframework==3.15.2      # Simple, Claude knows it well
django-ninja==1.3.0              # FastAPI-style endpoints
django-types==0.19.1             # Type stubs for Claude
djangorestframework-stubs==3.15.0
django-filter==24.3              # Easy filtering
drf-spectacular==0.27.2          # OpenAPI docs
django-cors-headers==4.3.1

# Database & Environment
psycopg2-binary
python-decouple==3.8             # Environment variables

# Medical specific
django-encrypted-model-fields==0.6.5
django-health-check              # Health monitoring and system status
django-auditlog==3.2.1
django-guardian==2.4.0           # Object-level permissions
hl7==0.4.5                       # HL7 message parsing for healthcare
pydicom==3.0.1                   # DICOM medical image handling

# Performance
django-redis==5.4.0
redis==5.0.8
django-cachalot==2.8.0           # ORM caching
django-silk==5.1.0               # Profiling

# Development
django-extensions==3.2.3
django-debug-toolbar==4.2.0
ipython==8.19.0

# Background tasks & Media
celery==5.4.0
Pillow

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
cryptography
